pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "A news aggregation agent using LangGraph, Tavily, and Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "tavily-python>=0.5.0",
    "feedparser>=6.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "firecrawl-py>=1.0.0",
    "newsapi-python>=0.2.7",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "pyyaml>=6.0.0",
    "praw>=7.7.0",
]

[project.scripts]
news-agent = "news_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/news_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

